Calculate the target before planning
The target state was being calculated at each planning state, allowing
effects to leech into next iterations of the plan via the target.

This now calculates the target state only at the beginning, when the
current state comes from the actual system state. While this is still
not perfect, as the target may end up containing stuff that we really
might not want in the final state (think a container re-start, where
we don't care about the final containerId), the target at least is
calculated from real data

Change-type: patch