Refactor Diff interface
Creating a new instance with `Diff.of` will now return a function to
calculate pending operations. This removes the public `patch` function as that
is not used by the planner and that is replaced by a readonly property
returning the result of patching a source object with the provided target.

Change-type: minor