Improve task types for testing
Task.of now returns the proper type (MethodTask or ActionTask) depending
on the input. This makes it easier to use these types as part of
testing. However, this comes with the tradeoff that tasks need to at
least defined an `effect` property or a `method` property to be type
checked correctly.

Change-type: patch