Add support for redirect tasks
A redirect task allows to define one or more targets for the planner to seek when
receiving a target. For instance, if the agent is a robot that
needs to get from A -> B but there is an obstacle in the path, then the
task could identify the obstacle and tell the planner to go from A -> C
and then from C -> B in order to reach the target.

Change-type: minor