Add `path` and `target` to instructions
Currently we calculate the action id using a hashing function from the
path, target and task id. This is not really useful as in reality what
we need to compare are applications of an action to a specific state.

The planner is doing that comparison at the moment by comparing the
actual state objects, which is a slow operation. It should really
compare ids generated from the action metadata and the state.

This is a first step towards that.

Change-type: minor