Move the method expansion check to findPlan
A method could be expanded to a an action and further methods, before if
one of the sub-method calls returned an empty array, that would cause
the parent method to fail. Now we only consider the method failed if one
of the expansion conditions failed or the method after the expansion is
finshed, didn't add any new actions to the plan.

Change-type: patch