Add Parallel task support.
Parallel tasks are the complement of Method tasks. While a method
returns a list of instructions to be executed in sequence, a parallel
task returns a list of instructions to be executed in parallel. When a
parallel task is found, the planner will create a 'fork' in the plan,
where sub-plans are executed in concurrent executions.

While actions may now be executed in parallel, the state object is still
shared, so this PR also provides utilities for conflict detection during
testing.

Change-type: minor