Improve testing tools to work with parallel plans
This improves the API of the plan builder to be simpler and less
confusing. The `end` method of the builder now returns a string.
The `simplified` method has been renamed to `stringify` and now returns
a string instead of an array. This should make it easier to visually
inspect differences betwen the expected and the generated plan

Change-type: minor