Create the `Effect` type
The `Effect` type is a mix between a promise, a function and an
observable. It provides a way to combine an async computation with a
sync version which "simulates" the effect that the async computation
will have on the underlying system. These types are also composable via
`map` and `flatMap`, which would allow building more complex effects
from simpler ones, however this is outside the scope of the current PR
(see PR #33 for a use of those ideas).

Change-type: minor