Add default conditions for `create` and `delete` tasks
When grounding the task, the framework will add some existence
pre-conditions depending on the operation. This avoids the user from
having to add them. The API now also makes either action or effect
optional when defining action tasks.

Whenever a `create` operation is defined, the Task API will create an
empty object before passing the value to the effect or action functions.

Whenever a `delete` operation is defined, the API will first execute the
effect and action functions and then proceed to delete the referenced
object, after the functions succeed.

Change-type: minor