Rename Operation.value to Operation.target
This naming is more consistent with the rest of the API. We'll also add
soon a `source` value to the update operation for debugging purposes.

This also cleans up some unused code in the operation module.

This is an internal interface so this is not a breaking change.

Change-type: patch