Make Agent maxRetries count failed plan runs
Before only failures to find a plan would count against the maxRetries
Agent options, which is really an error in design as that would not
leave to abort a failing plan.

This also changes maxRetries to default to Infinity, leaving maxRetries:
0 to mean that the agent will exit after a single try.

While this fixes some bad design decisions, this is a breaking change.

Change-type: major