Convert Path type into a branded type
This prevents the code from using simple string as paths, and ensures
Path operations are used when splitting and transforming paths. The path
constructor now allows passsing arrays as input, and path segments will
be escaped according to IETF RFC 6901, preventing issues if using
slashes as dictionary keys in the state object

The interface does not change for users.

Change-type: minor