Improve input types for task constructors
The `condition` and `method` operations now receive readonly objects in
the general case. If the method is `create` the compiler now will warn
when trying to assign a value to a View on Action methods

This is a major change as code that previously would compile will no
longer compile.

Change-type: major