Export deepEqual as part of the `utils` namespace
The `deepEqual` function is a key piece of how the distance to a target
is calculated, as it allows to compare between the current and target
state. This function is flexible to allow users to define their own
comparison operations by providing an `equals` function on their models.

This function may also come handy for users of the library, so we export
it as part of the `mahler/utils` namespace

Change-type: minor