Allow lens variables for Sensor constructor
This is a first step in allowing sensors to be created dynamically on
the Agent when a new matching property is created/removed. Think of the
case where we want a new sensor whenever a new object is created (like a
process) in order to track state changes of that sub-element of the
global state.

Change-type: minor