Update sensor subscription at runtime init
With dynamic sensors, the runtime now only subscribes to sensors whose
lenses match an existing property in the current state. If a property
does not exist, the sensor is not used.

Similarly, when the state changes, the runtime now checks for
subscriptions that are no longer applicable, or new sensors that
should be subscribed to and updates its internal state of subscriptions.

Change-type: minor