Simplify Sensor interface.
This updates the sensor interface to no longer require a state reference
as input, but simply return a subscribable for changes on a specific
path. This makes the interface simpler and moves the responsibility of
updating the state to sensor subscribers.

This is however a breaking change, even though the contructor interface
does not changes.

Change-type: major