Move state rollback to Agent runtime
The `observe` module would observe changes on a state and revert to the
previous state if an error occurred, however this had a flaw as the
rollback would revert the whole state object, causing problems if
concurrent changes were performed in parallel branches of the plan. This
moves the reversal to the Agent runtime, where the path information can
be used to revert, limiting the reverted changes only to the path that
the action modifies.

Change-type: patch